feature/BB-18 - problems testing pages
