+++ src/components/organisms/Header/component.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import * as styles from "./styles.module.scss";
-import { ReactComponent as Logo } from "src/assets/images/logo.svg";
+import logo from "src/assets/images/logo.svg";
 import burger from "src/assets/icons/burger.svg";
 import Btn from "../../../components/atoms/Btn";
 
@@ -10,7 +10,7 @@ function Header({ location }) {
     <header className={styles.header}>
       <div className={styles.container}>
         <Link to="/" className={styles.link}>
-          <Logo />
+          <img src={logo} alt="#" />
         </Link>
         <Link to="/catalog" className={styles.link}>
           <Btn
